src/requirements.txt
asyncpg
discord.py>=1.3.2
psutil
